Blur with no extra framebuffer copies.

+++ src/client/renderer/r_framebuffer.h
@@ -26,7 +26,7 @@
 r_framebuffer_t R_CreateFramebuffer(GLint width, GLint height, int32_t attachments);
 void R_ClearFramebuffer(r_framebuffer_t *framebuffer);
 void R_CopyFramebufferAttachment(const r_framebuffer_t *framebuffer, r_attachment_t attachment, GLuint *texture);
-void R_BlurFramebufferAttachment(r_framebuffer_t *framebuffer, r_attachment_t attachment, float blur);
+void R_BlurFramebufferAttachment(r_framebuffer_t *framebuffer, r_attachment_t attachment, int32_t blur);
 void R_BlitFramebufferAttachment(const r_framebuffer_t *framebuffer, r_attachment_t attachment, GLint x, GLint y, GLint w, GLint h);
 void R_BlitFramebuffer(const r_framebuffer_t *framebuffer, GLint x, GLint y, GLint w, GLint h);
 void R_ReadFramebufferAttachment(const r_framebuffer_t *framebuffer, r_attachment_t attachment, SDL_Surface **surface);
